+++ CMakeLists.txt
@@ -5,7 +5,7 @@ set(COMPONENT_ADD_INCLUDEDIRS
     src/lgfx/utility
     src/Fonts
     )
-file(GLOB UTIL_SRCS src/lgfx/utility/*.c src/Fonts/*.cpp)
-set(COMPONENT_SRCS ${UTIL_SRCS})
+file(GLOB SRCS src/lgfx/*.cpp src/lgfx/platforms/*.cpp src/lgfx/utility/*.c src/Fonts/*.cpp)
+set(COMPONENT_SRCS ${SRCS})
 
 register_component()
