+++ src/lgfx/platforms/samd51_common.cpp
@@ -1,0 +1,56 @@
+#if defined (__SAMD51__)
+
+#include "samd51_common.hpp"
+
+namespace lgfx
+{
+
+  void lgfxPinMode(std::uint32_t pin, pin_mode_t mode)
+  {
+    std::uint32_t port = pin>>8;
+    pin &= 0xFF;
+    std::uint32_t pinMask = (1ul << pin);
+
+    // Set pin mode according to chapter '22.6.3 I/O Pin Configuration'
+    switch ( mode )
+    {
+      case pin_mode_t::input:
+        // Set pin to input mode
+        PORT->Group[port].PINCFG[pin].reg=(std::uint8_t)(PORT_PINCFG_INEN) ;
+        PORT->Group[port].DIRCLR.reg = pinMask ;
+      break ;
+
+      case pin_mode_t::input_pullup:
+        // Set pin to input mode with pull-up resistor enabled
+        PORT->Group[port].PINCFG[pin].reg=(std::uint8_t)(PORT_PINCFG_INEN|PORT_PINCFG_PULLEN) ;
+        PORT->Group[port].DIRCLR.reg = pinMask ;
+
+        // Enable pull level (cf '22.6.3.2 Input Configuration' and '22.8.7 Data Output Value Set')
+        PORT->Group[port].OUTSET.reg = pinMask ;
+      break ;
+
+      case pin_mode_t::input_pulldown:
+        // Set pin to input mode with pull-down resistor enabled
+        PORT->Group[port].PINCFG[pin].reg=(std::uint8_t)(PORT_PINCFG_INEN|PORT_PINCFG_PULLEN) ;
+        PORT->Group[port].DIRCLR.reg = pinMask ;
+
+        // Enable pull level (cf '22.6.3.2 Input Configuration' and '22.8.6 Data Output Value Clear')
+        PORT->Group[port].OUTCLR.reg = pinMask ;
+      break ;
+
+      case pin_mode_t::output:
+        // enable input, to support reading back values, with pullups disabled
+        PORT->Group[port].PINCFG[pin].reg=(std::uint8_t)(PORT_PINCFG_INEN) ;
+
+        // Set pin to output mode
+        PORT->Group[port].DIRSET.reg = pinMask ;
+      break ;
+
+      default:
+        // do nothing
+      break ;
+    }
+  }
+}
+
+#endif
